Add files via upload
This is for the analysis of a growth delay assay that requires some specific experimental set up. 

This assumes you are using a 96-well plate with the outer wells unused (PBS to buffer evaporation) and cells plated in triplicate rows (condition 1: (B2:B11, C2:C11, D2:D11), condition 2: (E2:E11, F2:F11, G2:G11)). The X2 column wells are used as vehicle controls for their corresponding conditions and X3-X11 contain increasing concentrations of a treatment.

This is designed to work on my (James) laptop and so one would have to change the paths for it to work on their system. I annotated it pretty thoroughly, but if you have any questions please feel free to email me at jameselia777@gmail.com